added (') after pages in the sentence line 61
Before:
Power Pages inbuilt security is its core.
After:
Power Pages' inbuilt security is its core.

Explanation:
This is a minor grammatical error. Power Pages is a plural noun, and its security features are something it is possessive of, so I added an apostrophe to the end of the word.