Grammatical Corrections
There are minor grammatical errors that I have corrected to help with readability.

Line 31 - in this context, the preposition would be 'apply to' not 'apply for'.
Line 73 - on this line, the word necessary was missing an article, so I added 'the'. It is now consistent with the rest of the page.
Line 81 - the preposition was corrected, so 'in' was replaced with 'on'. I also added a comma after the word 'visit'.

when the English grammar is corrected, I found that web-based translators help keep the true meaning better.