Update configure-cdn.md
Added a note about custom certificates not supported when using CDN.
This info was received from MS Support:
"Unfortunately , there is no way to manage custom certificates when CDN is enabled. Because when CDN is enabled , the certificate will be managed by CDN team.
As there is no documentation as of today for this. On my request, our dev confirmed that he is going to get this documented tentatively in 2 weeks. I also informed that I'll share the document as soon as I get it from devs of product group."